requirements.txt
numpy>=1.0
opencv-python>=4.0
tensorflow>=2.0
matplotlib>=3.0
imghdr>=0.5
pandas>=1.0
flask>=1.0
flask-cors>=3.0
neo4j>=4.0
python-dotenv>=0.15